extracted parameter list shared by front and backend to common
