--- src/content/constants.tsx
@@ -1,120 +1,0 @@
-﻿export const intrinsicParameters = [
-  'a_1',
-  'a_2',
-  'chi_1_in_plane',
-  'chi_2_in_plane',
-  'chi_eff',
-  'chi_p',
-  'chirp_mass',
-  'chirp_mass_source',
-  'cos_tilt_1',
-  'cos_tilt_2',
-  'mass_1',
-  'mass_1_source',
-  'mass_2',
-  'mass_2_source',
-  'mass_ratio',
-  'phase',
-  'phi_1',
-  'phi_12',
-  'phi_2',
-  'phi_jl',
-  'psi',
-  'spin_1x',
-  'spin_1y',
-  'spin_1z',
-  'spin_2x',
-  'spin_2y',
-  'spin_2z',
-  'symmetric_mass_ratio',
-  'theta_jn',
-  'tilt_1',
-  'tilt_2',
-  'total_mass',
-  'total_mass_source'
-]
-
-export const extrinsicParameters = [
-  'H1_time',
-  'azimuth',
-  'comoving_distance',
-  'dec',
-  'geocent_time',
-  'iota',
-  'luminosity_distance',
-  'ra',
-  'redshift',
-  'zenith'
-]
-
-export const otherParameters = [
-  'H1_matched_filter_snr',
-  'H1_optimal_snr',
-  'L1_matched_filter_snr',
-  'L1_optimal_snr',
-  'log_likelihood',
-  'log_prior',
-  'minimum_frequency',
-  'recalib_H1_amplitude_0',
-  'recalib_H1_amplitude_1',
-  'recalib_H1_amplitude_2',
-  'recalib_H1_amplitude_3',
-  'recalib_H1_amplitude_4',
-  'recalib_H1_amplitude_5',
-  'recalib_H1_amplitude_6',
-  'recalib_H1_amplitude_7',
-  'recalib_H1_amplitude_8',
-  'recalib_H1_amplitude_9',
-  'recalib_H1_frequency_0',
-  'recalib_H1_frequency_1',
-  'recalib_H1_frequency_2',
-  'recalib_H1_frequency_3',
-  'recalib_H1_frequency_4',
-  'recalib_H1_frequency_5',
-  'recalib_H1_frequency_6',
-  'recalib_H1_frequency_7',
-  'recalib_H1_frequency_8',
-  'recalib_H1_frequency_9',
-  'recalib_H1_phase_0',
-  'recalib_H1_phase_1',
-  'recalib_H1_phase_2',
-  'recalib_H1_phase_3',
-  'recalib_H1_phase_4',
-  'recalib_H1_phase_5',
-  'recalib_H1_phase_6',
-  'recalib_H1_phase_7',
-  'recalib_H1_phase_8',
-  'recalib_H1_phase_9',
-  'recalib_L1_amplitude_0',
-  'recalib_L1_amplitude_1',
-  'recalib_L1_amplitude_2',
-  'recalib_L1_amplitude_3',
-  'recalib_L1_amplitude_4',
-  'recalib_L1_amplitude_5',
-  'recalib_L1_amplitude_6',
-  'recalib_L1_amplitude_7',
-  'recalib_L1_amplitude_8',
-  'recalib_L1_amplitude_9',
-  'recalib_L1_frequency_0',
-  'recalib_L1_frequency_1',
-  'recalib_L1_frequency_2',
-  'recalib_L1_frequency_3',
-  'recalib_L1_frequency_4',
-  'recalib_L1_frequency_5',
-  'recalib_L1_frequency_6',
-  'recalib_L1_frequency_7',
-  'recalib_L1_frequency_8',
-  'recalib_L1_frequency_9',
-  'recalib_L1_phase_0',
-  'recalib_L1_phase_1',
-  'recalib_L1_phase_2',
-  'recalib_L1_phase_3',
-  'recalib_L1_phase_4',
-  'recalib_L1_phase_5',
-  'recalib_L1_phase_6',
-  'recalib_L1_phase_7',
-  'recalib_L1_phase_8',
-  'recalib_L1_phase_9',
-  'reference_frequency',
-  'waveform_approximant'
-]
